Move connections to store in separate index files
