+++ src/components/Connector/index.js
@@ -1,0 +1,11 @@
+import {connect} from 'react-redux';
+import {setIsSeed} from "../../actions"
+import Connector from './Connector';
+
+const mapDispatchToProps = (dispatch) => {
+    return {
+        seedChange: (e) => dispatch(setIsSeed(e.target.checked))
+    }
+};
+
+export default connect(null, mapDispatchToProps)(Connector);
