+++ src/index.jsx
@@ -3,9 +3,9 @@ import ReactDOM from 'react-dom';
 import {Provider} from 'react-redux';
 import {createStore, applyMiddleware} from 'redux';
 import reducers from './reducers';
-import App from './components/App/App' ;
-import './index.styl';
+import App from './components/App' ;
 import logger from "./middleware/logger";
+import './index.styl';
 
 const store = createStore(reducers, applyMiddleware(logger));
